+++ application/app/src/main/AndroidManifest.xml
@@ -42,7 +42,7 @@
         <activity
             android:name=".select.plant.SelectPlantActivity"
             android:exported="true"
-            android:theme="@style/Theme.AppCompat.DayNight.NoActionBar">
+            android:theme="@style/Theme.Priend">
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -69,7 +69,7 @@
         <activity
             android:name=".main.MainActivity"
             android:exported="false"
-            android:theme="@style/Theme.AppCompat.DayNight.NoActionBar">
+            android:theme="@style/Theme.Priend">
 
 
         </activity>
