+++ application/app/src/main/java/com/example/priend/main/info/view/list/recyclerview/InfoItemType.kt
@@ -1,0 +1,6 @@
+package com.example.priend.main.info.view.list.recyclerview
+
+enum class InfoItemType {
+    ADD,
+    PLANT
+}
