+++ application/app/src/main/java/com/example/priend/main/info/view/list/recyclerview/viewholder/InfoPlantViewHolder.kt
@@ -1,16 +1,15 @@
-package com.example.priend.main.info.view.recyclerview.viewholder
+package com.example.priend.main.info.view.list.recyclerview.viewholder
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import com.example.priend.common.view.RecyclerViewItemClickListener
-import com.example.priend.databinding.ItemInfoAddBinding
-import com.example.priend.main.info.view.recyclerview.InfoItem
+import com.example.priend.databinding.ItemInfoBinding
+import com.example.priend.main.info.view.list.recyclerview.InfoItem
 
-class InfoAddViewHolder(
-    private val binding: ItemInfoAddBinding,
+class InfoPlantViewHolder(
+    private val binding: ItemInfoBinding,
     private val clickListener: RecyclerViewItemClickListener
 ) : RecyclerView.ViewHolder(binding.root), View.OnClickListener, InfoViewHolder {
-
     init {
         binding.root.setOnClickListener(this)
     }
@@ -23,11 +22,11 @@ class InfoAddViewHolder(
     }
 
     override fun bind(item: InfoItem) {
-        val position = adapterPosition
-        if (position != RecyclerView.NO_POSITION) {
-            clickListener.onItemClick(position)
+        binding.apply {
+            itemInfoPlantName.text = item.title
+            itemInfoTemperature.text = item.temperature
+            itemInfoStartDate.text = item.startDate
+            itemInfoSoilMoisture.text = item.soilMoisture
         }
     }
-
-
 }
