--- application/app/src/main/java/com/example/priend/main/info/view/recyclerview/InfoItemType.kt
@@ -1,6 +1,0 @@
-package com.example.priend.main.info.view.recyclerview
-
-enum class InfoItemType {
-    ADD,
-    PLANT
-}
