--- application/app/src/main/java/com/example/priend/main/info/view/recyclerview/viewholder/InfoViewHolder.kt
@@ -1,7 +1,0 @@
-package com.example.priend.main.info.view.recyclerview.viewholder
-
-import com.example.priend.main.info.view.recyclerview.InfoItem
-
-interface InfoViewHolder {
-    fun bind(item: InfoItem)
-}
