+++ application/app/src/main/res/drawable/menu_selector_color.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/teal_700" android:state_checked="true" />
+    <item android:color="#CCCCCC" android:state_checked="false" />
+
+</selector>
