+++ application/app/src/main/res/layout/activity_main.xml
@@ -25,6 +25,9 @@
         android:layout_height="wrap_content"
         app:layout_constraintTop_toBottomOf="@+id/navi_host_fragment"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:itemIconTint="@drawable/menu_selector_color"
+        app:itemTextColor="@drawable/menu_selector_color"
+        android:background="@color/white"
         app:menu="@menu/nav_menu"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
