+++ application/app/src/main/res/layout/fragment_info.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".main.info.view.InfoFragment">
+    tools:context=".main.info.view.list.InfoFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -21,12 +21,12 @@
                 android:layout_height="match_parent"
                 android:src="@drawable/ic_priend"
                 app:layout_constraintHorizontal_weight="0.5"
-                app:layout_constraintEnd_toStartOf="@+id/info_title_text"
+                app:layout_constraintEnd_toStartOf="@+id/info_detail_plant_name"
                 app:layout_constraintStart_toStartOf="parent"
                 tools:layout_editor_absoluteY="10dp" />
 
             <TextView
-                android:id="@+id/info_title_text"
+                android:id="@+id/info_detail_plant_name"
                 android:layout_width="0dp"
                 android:layout_height="match_parent"
                 android:layout_marginStart="10sp"
@@ -43,7 +43,7 @@
                 android:id="@+id/info_menu_icon"
                 android:layout_width="0dp"
                 android:layout_height="match_parent"
-                app:layout_constraintStart_toEndOf="@id/info_title_text"
+                app:layout_constraintStart_toEndOf="@id/info_detail_plant_name"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_weight="0.3"
                 android:src="@drawable/ic_menu"/>
